Consolidate similar serializers
