+++ app/serializers/garden_plant_serializer.rb
@@ -1,23 +1,23 @@
-class GardenPlantSerializer
-  include JSONAPI::Serializer
-  attributes  :id,
-              :name,
-              :plant_type,
-              :days_relative_to_frost_date,
-              :days_to_maturity,
-              :hybrid_status,
-              :organic
-
-
-  def self.error(message)
-    {
-      "error": message
-    }
-  end
-
-  def self.confirm
-    {
-      "status": "success"
-    }
-  end
-end
+# class GardenPlantSerializer
+#   include JSONAPI::Serializer
+#   attributes  :id,
+#               :name,
+#               :plant_type,
+#               :days_relative_to_frost_date,
+#               :days_to_maturity,
+#               :hybrid_status,
+#               :organic
+#
+#
+#   def self.error(message)
+#     {
+#       "error": message
+#     }
+#   end
+#
+#   def self.confirm
+#     {
+#       "status": "success"
+#     }
+#   end
+# end
